+++ packages/contracts-core/test/suite/Destination.t.sol
@@ -1,7 +1,8 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.17;
 
-import {CallerNotInbox, NotaryInDispute, OutdatedNonce} from "../../contracts/libs/Errors.sol";
+import {DISPUTE_TIMEOUT_NOTARY} from "../../contracts/libs/Constants.sol";
+import {CallerNotInbox, DisputeTimeoutNotOver, NotaryInDispute, OutdatedNonce} from "../../contracts/libs/Errors.sol";
 import {SNAPSHOT_MAX_STATES} from "../../contracts/libs/memory/Snapshot.sol";
 import {DisputeFlag} from "../../contracts/libs/Structures.sol";
 import {IAgentSecured} from "../../contracts/interfaces/IAgentSecured.sol";
@@ -40,8 +41,7 @@ contract DestinationTest is ExecutionHubTest {
         // Check version
         assertEq(dst.version(), LATEST_VERSION, "!version");
         // Check pending Agent Merkle Root
-        (bool rootPassed, bool rootPending) = dst.passAgentRoot();
-        assertFalse(rootPassed);
+        (bool rootPending) = dst.passAgentRoot();
         assertFalse(rootPending);
     }
 
@@ -194,6 +194,46 @@ contract DestinationTest is ExecutionHubTest {
         InterfaceDestination(localDestination()).acceptAttestation(agentIndex[notary], 0, "", 0, new ChainGas[](0));
     }
 
+    function test_acceptAttestation_revert_notaryWonDisputeTimeout() public {
+        address notary = domains[DOMAIN_LOCAL].agent;
+        address guard = domains[0].agent;
+
+        Random memory random = Random("salt");
+        RawSnapshot memory rawSnap = random.nextSnapshot();
+        RawAttestation memory ra = random.nextAttestation({rawSnap: rawSnap, nonce: 1});
+        uint256[] memory snapGas = rawSnap.snapGas();
+        (bytes memory attPayload, bytes memory attSig) = signAttestation(notary, ra);
+
+        openTestDispute({guardIndex: agentIndex[guard], notaryIndex: agentIndex[notary]});
+        skip(7 days);
+        resolveTestDispute({slashedIndex: agentIndex[guard], rivalIndex: agentIndex[notary]});
+        skip(DISPUTE_TIMEOUT_NOTARY - 1);
+        vm.expectRevert(DisputeTimeoutNotOver.selector);
+        lightInbox.submitAttestation(attPayload, attSig, ra._agentRoot, snapGas);
+    }
+
+    function test_acceptAttestation_afterNotaryDisputeTimeout() public {
+        address notary = domains[DOMAIN_LOCAL].agent;
+        address guard = domains[0].agent;
+
+        Random memory random = Random("salt");
+        RawSnapshot memory rawSnap = random.nextSnapshot();
+        bytes32 snapRoot = rawSnap.castToSnapshot().calculateRoot();
+        // Sanity check
+        assert(testedEH().getAttestationNonce(snapRoot) == 0);
+        RawAttestation memory ra = random.nextAttestation({rawSnap: rawSnap, nonce: 1});
+        uint256[] memory snapGas = rawSnap.snapGas();
+        (bytes memory attPayload, bytes memory attSig) = signAttestation(notary, ra);
+
+        openTestDispute({guardIndex: agentIndex[guard], notaryIndex: agentIndex[notary]});
+        skip(7 days);
+        resolveTestDispute({slashedIndex: agentIndex[guard], rivalIndex: agentIndex[notary]});
+        skip(DISPUTE_TIMEOUT_NOTARY);
+        lightInbox.submitAttestation(attPayload, attSig, ra._agentRoot, snapGas);
+        // Snapshot root should be known to ExecutionHub (Destination)
+        assertGt(testedEH().getAttestationNonce(snapRoot), 0);
+    }
+
     function test_acceptAttestation_revert_lowerNonce() public {
         Random memory random = Random("salt");
         RawAttestation memory firstRA = random.nextAttestation({nonce: 2});
@@ -221,22 +261,34 @@ contract DestinationTest is ExecutionHubTest {
         lightInbox.submitAttestation(attPayload, attSig, ra._agentRoot, snapGas);
     }
 
-    function test_acceptAttestation_notAccepted_agentRootUpdated(
+    function test_acceptAttestation_passesAgentRoot(
         RawAttestation memory firstRA,
+        RawAttestation memory secondRA,
         uint32 firstRootSubmittedAt
     ) public {
         bytes32 agentRootLM = lightManager.agentRoot();
         vm.assume(firstRA._agentRoot != agentRootLM);
+        vm.assume(firstRA.snapRoot != secondRA.snapRoot);
+        vm.assume(secondRA.nonce > firstRA.nonce);
         test_submitAttestation(firstRA, firstRootSubmittedAt);
         skip(AGENT_ROOT_OPTIMISTIC_PERIOD);
-        // Mock a call from lightInbox, could as well use the empty values as they won't be checked for validity
-        vm.prank(address(lightInbox));
-        assertFalse(InterfaceDestination(localDestination()).acceptAttestation(0, 0, "", 0, new ChainGas[](0)));
+        // Form a second attestation: Notary 1
+        RawSnapshot memory rs = Random(secondRA.snapRoot).nextSnapshot();
+        secondRA._snapGasHash = rs.snapGasHash();
+        secondRA.setDataHash();
+        uint256[] memory snapGas = rs.snapGas();
+        address notaryS = domains[DOMAIN_LOCAL].agents[1];
+        (bytes memory attPayload, bytes memory attSig) = signAttestation(notaryS, secondRA);
+        uint256 newRootTimestamp = block.timestamp;
+        assertTrue(lightInbox.submitAttestation(attPayload, attSig, secondRA._agentRoot, snapGas));
         (uint40 snapRootTime, uint40 agentRootTime, uint32 index) =
             InterfaceDestination(localDestination()).destStatus();
-        assertEq(snapRootTime, firstRootSubmittedAt);
-        assertEq(agentRootTime, firstRootSubmittedAt);
-        assertEq(index, agentIndex[domains[DOMAIN_LOCAL].agent]);
+        // Dest status should point to the new root
+        assertEq(snapRootTime, newRootTimestamp);
+        assertEq(agentRootTime, newRootTimestamp);
+        assertEq(index, agentIndex[notaryS]);
+        // New agent root should be pending in Destination
+        assertEq(InterfaceDestination(localDestination()).nextAgentRoot(), secondRA._agentRoot);
         // Should update the Agent Merkle Root
         assertEq(lightManager.agentRoot(), firstRA._agentRoot);
     }
@@ -280,8 +332,7 @@ contract DestinationTest is ExecutionHubTest {
         test_submitAttestation_updatesAgentRoot(ra, rootSubmittedAt);
         timePassed = timePassed % AGENT_ROOT_OPTIMISTIC_PERIOD;
         skip(timePassed);
-        (bool rootPassed, bool rootPending) = InterfaceDestination(localDestination()).passAgentRoot();
-        assertFalse(rootPassed);
+        bool rootPending = InterfaceDestination(localDestination()).passAgentRoot();
         assertTrue(rootPending);
         assertEq(lightManager.agentRoot(), agentRootLM);
     }
@@ -290,10 +341,121 @@ contract DestinationTest is ExecutionHubTest {
         // Submit attestation that updates `nextAgentRoot`
         test_submitAttestation_updatesAgentRoot(ra, rootSubmittedAt);
         skip(AGENT_ROOT_OPTIMISTIC_PERIOD);
-        (bool rootPassed, bool rootPending) = InterfaceDestination(localDestination()).passAgentRoot();
-        assertTrue(rootPassed);
+        bool rootPending = InterfaceDestination(localDestination()).passAgentRoot();
         assertFalse(rootPending);
         assertEq(lightManager.agentRoot(), ra._agentRoot);
+    }
+
+    /// @dev Submits a mock attestation and opens a dispute for the Notary that signed it.
+    function prepareAgentRootDisputeTest() internal returns (bytes32 newAgentRoot) {
+        address notary = domains[DOMAIN_LOCAL].agent;
+        address guard = domains[0].agent;
+
+        Random memory random = Random("salt");
+        RawSnapshot memory rawSnap = random.nextSnapshot();
+        RawAttestation memory ra = random.nextAttestation({rawSnap: rawSnap, nonce: 1});
+        newAgentRoot = ra._agentRoot;
+        uint256[] memory snapGas = rawSnap.snapGas();
+
+        (bytes memory attPayload, bytes memory attSig) = signAttestation(notary, ra);
+        lightInbox.submitAttestation(attPayload, attSig, ra._agentRoot, snapGas);
+        // Sanity check
+        assert(InterfaceDestination(localDestination()).nextAgentRoot() == newAgentRoot);
+        openTestDispute({guardIndex: agentIndex[guard], notaryIndex: agentIndex[notary]});
+    }
+
+    /// @dev Resolves test dispute above in favor of the Notary.
+    function prepareNotaryWonDisputeTest() internal {
+        address notary = domains[DOMAIN_LOCAL].agent;
+        address guard = domains[0].agent;
+        resolveTestDispute({slashedIndex: agentIndex[guard], rivalIndex: agentIndex[notary]});
+    }
+
+    function test_passAgentRoot_notaryInDispute_optimisticPeriodNotOver() public {
+        bytes32 oldRoot = lightManager.agentRoot();
+        prepareAgentRootDisputeTest();
+        skip(AGENT_ROOT_OPTIMISTIC_PERIOD - 1);
+        bool rootPending = InterfaceDestination(localDestination()).passAgentRoot();
+        // Should not pass the root
+        assertEq(lightManager.agentRoot(), oldRoot);
+        // Should clear pending
+        assertFalse(rootPending);
+        assertEq(InterfaceDestination(localDestination()).nextAgentRoot(), oldRoot);
+    }
+
+    function test_passAgentRoot_notaryInDispute_optimisticPeriodOver() public {
+        bytes32 oldRoot = lightManager.agentRoot();
+        prepareAgentRootDisputeTest();
+        skip(AGENT_ROOT_OPTIMISTIC_PERIOD);
+        bool rootPending = InterfaceDestination(localDestination()).passAgentRoot();
+        // Should not pass the root
+        assertEq(lightManager.agentRoot(), oldRoot);
+        // Should clear pending
+        assertFalse(rootPending);
+        assertEq(InterfaceDestination(localDestination()).nextAgentRoot(), oldRoot);
+    }
+
+    function test_passAgentRoot_notaryWonDisputeTimeout_optimisticPeriodNotOver() public {
+        bytes32 oldRoot = lightManager.agentRoot();
+        bytes32 newRoot = prepareAgentRootDisputeTest();
+        skip(AGENT_ROOT_OPTIMISTIC_PERIOD - DISPUTE_TIMEOUT_NOTARY);
+        prepareNotaryWonDisputeTest();
+        skip(DISPUTE_TIMEOUT_NOTARY - 1);
+        // Time since attestation was submitted: AGENT_ROOT_OPTIMISTIC_PERIOD - 1
+        // Time sinceNotary won the dispute: DISPUTE_TIMEOUT_NOTARY - 1
+        bool rootPending = InterfaceDestination(localDestination()).passAgentRoot();
+        // Should not pass the root
+        assertEq(lightManager.agentRoot(), oldRoot);
+        // Should not clear pending
+        assertTrue(rootPending);
+        assertEq(InterfaceDestination(localDestination()).nextAgentRoot(), newRoot);
+    }
+
+    function test_passAgentRoot_notaryWonDisputeTimeout_optimisticPeriodOver() public {
+        bytes32 oldRoot = lightManager.agentRoot();
+        bytes32 newRoot = prepareAgentRootDisputeTest();
+        skip(AGENT_ROOT_OPTIMISTIC_PERIOD - (DISPUTE_TIMEOUT_NOTARY - 1));
+        prepareNotaryWonDisputeTest();
+        skip(DISPUTE_TIMEOUT_NOTARY - 1);
+        // Time since attestation was submitted: AGENT_ROOT_OPTIMISTIC_PERIOD
+        // Time sinceNotary won the dispute: DISPUTE_TIMEOUT_NOTARY - 1
+        bool rootPending = InterfaceDestination(localDestination()).passAgentRoot();
+        // Should not pass the root
+        assertEq(lightManager.agentRoot(), oldRoot);
+        // Should not clear pending
+        assertTrue(rootPending);
+        assertEq(InterfaceDestination(localDestination()).nextAgentRoot(), newRoot);
+    }
+
+    function test_passAgentRoot_afterNotaryDisputeTimeout_optimisticPeriodNotOver() public {
+        bytes32 oldRoot = lightManager.agentRoot();
+        bytes32 newRoot = prepareAgentRootDisputeTest();
+        skip(AGENT_ROOT_OPTIMISTIC_PERIOD - 1 - DISPUTE_TIMEOUT_NOTARY);
+        prepareNotaryWonDisputeTest();
+        skip(DISPUTE_TIMEOUT_NOTARY);
+        // Time since attestation was submitted: AGENT_ROOT_OPTIMISTIC_PERIOD - 1
+        // Time sinceNotary won the dispute: DISPUTE_TIMEOUT_NOTARY
+        bool rootPending = InterfaceDestination(localDestination()).passAgentRoot();
+        // Should not pass the root
+        assertEq(lightManager.agentRoot(), oldRoot);
+        // Should not clear pending
+        assertTrue(rootPending);
+        assertEq(InterfaceDestination(localDestination()).nextAgentRoot(), newRoot);
+    }
+
+    function test_passAgentRoot_afterNotaryDisputeTimeout_optimisticPeriodOver() public {
+        bytes32 newRoot = prepareAgentRootDisputeTest();
+        skip(AGENT_ROOT_OPTIMISTIC_PERIOD - DISPUTE_TIMEOUT_NOTARY);
+        prepareNotaryWonDisputeTest();
+        skip(DISPUTE_TIMEOUT_NOTARY);
+        // Time since attestation was submitted: AGENT_ROOT_OPTIMISTIC_PERIOD
+        // Time sinceNotary won the dispute: DISPUTE_TIMEOUT_NOTARY
+        bool rootPending = InterfaceDestination(localDestination()).passAgentRoot();
+        // Should pass the root
+        assertEq(lightManager.agentRoot(), newRoot);
+        // Should clear pending
+        assertFalse(rootPending);
+        assertEq(InterfaceDestination(localDestination()).nextAgentRoot(), newRoot);
     }
 
     // ═════════════════════════════════════════════════ GAS DATA ══════════════════════════════════════════════════════
@@ -387,20 +549,36 @@ contract DestinationTest is ExecutionHubTest {
         assertEq(dataMaturity, 0);
     }
 
-    function test_getGasData_notaryInDispute(Random memory random) public {
-        RawSnapshot memory firstSnap;
-        firstSnap.states = new RawState[](2);
-        firstSnap.states[0] = random.nextState({origin: DOMAIN_REMOTE, nonce: random.nextUint32()});
-        firstSnap.states[1] = random.nextState({origin: DOMAIN_SYNAPSE, nonce: random.nextUint32()});
-        RawAttestation memory ra = random.nextAttestation(firstSnap, random.nextUint32());
-        address firstNotary = domains[DOMAIN_LOCAL].agents[0];
-        (bytes memory attPayload, bytes memory attSig) = signAttestation(firstNotary, ra);
-        uint256[] memory firstSnapGas = firstSnap.snapGas();
-        // Submit first attestation
-        lightInbox.submitAttestation(attPayload, attSig, ra._agentRoot, firstSnapGas);
-        skip(random.nextUint32());
-        // Open dispute
-        openDispute({guard: domains[0].agent, notary: firstNotary});
+    function prepareGasDataDisputeTest() internal returns (GasData remoteGasData, GasData synapseGasData) {
+        address notary = domains[DOMAIN_LOCAL].agent;
+        address guard = domains[0].agent;
+
+        Random memory random = Random("salt");
+        RawSnapshot memory rawSnap = RawSnapshot(new RawState[](2));
+        rawSnap.states[0] = random.nextState({origin: DOMAIN_REMOTE, nonce: 1});
+        rawSnap.states[1] = random.nextState({origin: DOMAIN_SYNAPSE, nonce: 2});
+        remoteGasData = rawSnap.states[0].castToState().gasData();
+        synapseGasData = rawSnap.states[1].castToState().gasData();
+
+        RawAttestation memory ra = random.nextAttestation({rawSnap: rawSnap, nonce: 1});
+        uint256[] memory snapGas = rawSnap.snapGas();
+        (bytes memory attPayload, bytes memory attSig) = signAttestation(notary, ra);
+        lightInbox.submitAttestation(attPayload, attSig, ra._agentRoot, snapGas);
+        // Sanity checks
+        {
+            (GasData gasData,) = InterfaceDestination(localDestination()).getGasData(DOMAIN_REMOTE);
+            assert(GasData.unwrap(gasData) == GasData.unwrap(remoteGasData));
+        }
+        {
+            (GasData gasData,) = InterfaceDestination(localDestination()).getGasData(DOMAIN_SYNAPSE);
+            assert(GasData.unwrap(gasData) == GasData.unwrap(synapseGasData));
+        }
+        openTestDispute({guardIndex: agentIndex[guard], notaryIndex: agentIndex[notary]});
+    }
+
+    function test_getGasData_notaryInDispute() public {
+        prepareGasDataDisputeTest();
+        skip(7 days);
         // Check getGasData
         (GasData gasData, uint256 dataMaturity) = InterfaceDestination(localDestination()).getGasData(DOMAIN_REMOTE);
         assertEq(GasData.unwrap(gasData), 0);
@@ -408,6 +586,34 @@ contract DestinationTest is ExecutionHubTest {
         (gasData, dataMaturity) = InterfaceDestination(localDestination()).getGasData(DOMAIN_SYNAPSE);
         assertEq(GasData.unwrap(gasData), 0);
         assertEq(dataMaturity, 0);
+    }
+
+    function test_getGasData_notaryWonDisputeTimeout() public {
+        prepareGasDataDisputeTest();
+        skip(7 days);
+        prepareNotaryWonDisputeTest();
+        skip(DISPUTE_TIMEOUT_NOTARY - 1);
+        // Check getGasData
+        (GasData gasData, uint256 dataMaturity) = InterfaceDestination(localDestination()).getGasData(DOMAIN_REMOTE);
+        assertEq(GasData.unwrap(gasData), 0);
+        assertEq(dataMaturity, 0);
+        (gasData, dataMaturity) = InterfaceDestination(localDestination()).getGasData(DOMAIN_SYNAPSE);
+        assertEq(GasData.unwrap(gasData), 0);
+        assertEq(dataMaturity, 0);
+    }
+
+    function test_getGasData_afterNotaryDisputeTimeout() public {
+        (GasData remoteGasData, GasData synapseGasData) = prepareGasDataDisputeTest();
+        skip(7 days);
+        prepareNotaryWonDisputeTest();
+        skip(DISPUTE_TIMEOUT_NOTARY);
+        // Check getGasData
+        (GasData gasData, uint256 dataMaturity) = InterfaceDestination(localDestination()).getGasData(DOMAIN_REMOTE);
+        assertEq(GasData.unwrap(gasData), GasData.unwrap(remoteGasData));
+        assertEq(dataMaturity, 7 days + DISPUTE_TIMEOUT_NOTARY);
+        (gasData, dataMaturity) = InterfaceDestination(localDestination()).getGasData(DOMAIN_SYNAPSE);
+        assertEq(GasData.unwrap(gasData), GasData.unwrap(synapseGasData));
+        assertEq(dataMaturity, 7 days + DISPUTE_TIMEOUT_NOTARY);
     }
 
     // ══════════════════════════════════════════════════ HELPERS ══════════════════════════════════════════════════════