+++ services/sinner/db/model/model.go
@@ -44,7 +44,7 @@ var (
 // MessageStatus is the table holding the status of each message.
 type MessageStatus struct {
 	// MessageHash is the message hash.
-	MessageHash string `gorm:"column:message_hash;uniqueIndex:idx_message_hash_status"`
+	MessageHash string `gorm:"type:varchar(66);column:message_hash;primaryKey"`
 	// OriginTxHash is the txhash when the origin event was emitted.
 	OriginTxHash string `gorm:"column:origin_txhash"`
 	// DestinationTxHash is the txhash when the destination event was emitted.
@@ -72,7 +72,7 @@ type OriginSent struct {
 	// MessageID is the keccaked message.
 	MessageID string `gorm:"column:message_id"`
 	// MessageHash is the message hash.
-	MessageHash string `gorm:"column:message_hash;primaryKey"`
+	MessageHash string `gorm:"type:varchar(66);column:message_hash;primaryKey"`
 	// ChainID is the chain id.
 	ChainID uint32 `gorm:"column:chain_id;primaryKey;index:idx_tx_hash_origin,priority:2,sort:desc"`
 	// Destination is the destination chain id.
@@ -114,7 +114,7 @@ type Executed struct {
 	// TxIndex is the index of the tx in a block.
 	TxIndex uint `gorm:"column:tx_index"`
 	// MessageHash is the message hash.
-	MessageHash string `gorm:"column:message_hash;primaryKey"`
+	MessageHash string `gorm:"type:varchar(66);column:message_hash;primaryKey"`
 	// ChainID is the chain id.
 	ChainID uint32 `gorm:"column:chain_id;primaryKey;index:idx_tx_hash_executed,priority:2,sort:desc"`
 	// RemoteDomain is the destination.