+++ packages/synapse-interface/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.163](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.162...@synapsecns/synapse-interface@0.1.163) (2023-11-02)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.162](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.161...@synapsecns/synapse-interface@0.1.162) (2023-11-01)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.161](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.160...@synapsecns/synapse-interface@0.1.161) (2023-11-01)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.160](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.159...@synapsecns/synapse-interface@0.1.160) (2023-11-01)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
 ## [0.1.159](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.158...@synapsecns/synapse-interface@0.1.159) (2023-10-27)
 
 **Note:** Version bump only for package @synapsecns/synapse-interface